Summary:Tweakings in the python code for database intialisation and prediction
