+++ requirements.txt
@@ -1,18 +1,32 @@
 absl-py==2.1.0
+annotated-types==0.7.0
 astunparse==1.6.3
 certifi==2024.8.30
 charset-normalizer==3.4.0
+click==8.1.7
+contourpy==1.3.0
+cycler==0.12.1
+docker-pycreds==0.4.0
+eval_type_backport==0.2.0
 flatbuffers==24.3.25
+fonttools==4.55.3
 gast==0.6.0
+gitdb==4.0.11
+GitPython==3.1.43
 google-pasta==0.2.0
 grpcio==1.68.0
 h5py==3.12.1
 idna==3.10
+importlib_metadata==8.5.0
+importlib_resources==6.4.5
+joblib==1.4.2
 keras==3.6.0
+kiwisolver==1.4.7
 libclang==18.1.1
 Markdown==3.7
 markdown-it-py==3.0.0
 MarkupSafe==3.0.2
+matplotlib==3.9.4
 mdurl==0.1.2
 ml-dtypes==0.3.2
 namex==0.0.8
@@ -20,18 +34,34 @@ numpy==1.26.4
 opt_einsum==3.4.0
 optree==0.13.1
 packaging==24.2
+pillow==11.0.0
+platformdirs==4.3.6
 protobuf==4.25.5
+psutil==6.1.0
+psycopg2-binary==2.9.10
+pydantic==2.10.3
+pydantic_core==2.27.1
 Pygments==2.18.0
+pyparsing==3.2.0
+python-dateutil==2.9.0.post0
+PyYAML==6.0.2
 requests==2.32.3
 rich==13.9.4
-setuptools==75.5.0
+scikit-learn==1.6.0
+scipy==1.13.1
+sentry-sdk==2.19.2
+setproctitle==1.3.4
 six==1.16.0
+smmap==5.0.1
 tensorboard==2.16.2
 tensorboard-data-server==0.7.2
 tensorflow==2.16.2
+tensorflow-io-gcs-filesystem==0.37.1
 termcolor==2.5.0
+threadpoolctl==3.5.0
 typing_extensions==4.12.2
 urllib3==2.2.3
+wandb==0.19.1
 Werkzeug==3.1.3
-wheel==0.45.0
 wrapt==1.16.0
+zipp==3.21.0
